+++ src/clj_kondo/impl/config.clj
@@ -96,6 +96,7 @@
                       #_[clojure.test]}
               :refer-all {:level :warning
                           :exclude #{}}
+              :refer-clojure-exclude-non-existing-var {:level :warning}
               :use {:level :warning}
               :missing-else-branch {:level :warning}
               :if-nil-return {:level :off}